news-video-app/requirements.txt
# Core Dependencies
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# AI & NLP
openai==1.3.7
anthropic==0.8.1
langchain==0.1.0

# Text-to-Speech
gtts==2.4.0
pydub==0.25.1
edge-tts==6.1.9

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Video Processing
moviepy==1.0.3
# FFmpeg required (install separately)

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Web Scraping
selenium==4.16.0
playwright==1.40.0
fake-useragent==1.4.0

# Scheduler
APScheduler==3.10.4

# Image Search & Download
unsplash-python==1.1.0
pexels-api==1.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
colorlog==6.8.0

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
